Philips color swap now in init function
There's an instruction in the Philips driver to swap RGB/BGR (MADCTL).
Awesome. Now, instead of calling swapColors for every setPixel we can
set it up in the display's initialization.

This swap can be set with a new (optional) parameter in init function,
but should default to the correct value.

Other changes:
* Better looking init function. Now only runs instructions specific to
driver type.
* Modified how contrast was set for Philips/Epson.
* Defaulting all example code to PHILIPS, this looks to be the most
common display now.
* Minor fixes in ChronoLCD code.